+++ packages/editor/components/CMSRoot/messages.jsx
@@ -1,0 +1,6 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  hasResources: 'This app has the following resources: ',
+  noResources: 'This app has no resources.',
+});