+++ packages/editor/components/NavLink/index.jsx
@@ -1,0 +1,5 @@
+import { withRouter } from 'react-router-dom';
+
+import NavLink from './NavLink';
+
+export default withRouter(NavLink);