+++ packages/editor/components/ResourceTable/index.jsx
@@ -1,0 +1,19 @@
+import { injectIntl } from 'react-intl';
+import { connect } from 'react-redux';
+
+import { push } from '../../actions/message';
+import ResourceTable from './ResourceTable';
+
+function mapStateToProps(state, ownProps) {
+  return {
+    app: state.apps.apps.find(app => app.id === Number(ownProps.match.params.id)),
+    resourceName: ownProps.match.params.resourceName,
+  };
+}
+
+export default injectIntl(
+  connect(
+    mapStateToProps,
+    { push },
+  )(ResourceTable),
+);