+++ packages/editor/templates/index.js
@@ -1,5 +1,6 @@
 import Empty from './empty';
 import PersonApp from './person';
 import HolidayApp from './holiday';
+import UnlitteredApp from './unlittered';
 
-export default [Empty, PersonApp, HolidayApp];
+export default [Empty, PersonApp, HolidayApp, UnlitteredApp];