+++ packages/server/utils/sentryDsnToReportUri.js
@@ -1,9 +1,7 @@
-const regex = /(https?):\/\/([\w]+)@([\w.]+)\/(\d+)/;
-
 export default function sentryDsnToReportUri(dsn) {
   if (!dsn) {
     return null;
   }
-  const [, protocol, pubKey, host, id] = dsn.match(regex);
-  return `${protocol}://${host}/api/${id}/security/?sentry_key=${pubKey}`;
+  const { protocol, username, host, pathname } = new URL(dsn);
+  return `${protocol}//${host}/api${pathname}/security/?sentry_key=${username}`;
 }