+++ CHANGELOG.md
@@ -7,20 +7,27 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.5.0] - 2019-04-11
+
 ### Added
 
 - App: Add support for app description field.
 - App: Add ´resource´ action type. This further simplifies the dataflow for retrieving and creating
   data.
+- App: Add support for building query strings for `resource` and `request` actions.
 - Block: Add List block.
+- Block: Add Markdown block.
 - CLI: Add support for authentication.
 - Editor: Add support for app description field.
+- Editor: Add `Person App`, `Holiday App`, and `Unlittered App` app templates.
 - Server: Add support for app description field.
 
 ### Changed
 
 - App: Rename ´definitions´ to ´resources´ in app definition.
 - Block: Rewrite and simplify detail-viewer block.
+- Block: Replace usage of resource in map block with a load action.
+- Server: Add security scopes for API calls.
 
 ## [0.4.2] - 2019-02-26